Update user rights enumeration function for UserRightsUtil

+++ UserRightsUtil/UserRightsUtil/Interop/NativeMethods.cs
@@ -55,7 +55,7 @@ namespace UserRightsUtil.Interop
             IntPtr PolicyHandle,
             IntPtr pSID,
             out IntPtr UserRights, // LSA_UNICODE_STRING[]
-            out ulong CountOfRights);
+            out uint CountOfRights);
 
         [DllImport("advapi32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
         public static extern int LsaEnumerateAccountsWithUserRight(
@@ -118,7 +118,7 @@ namespace UserRightsUtil.Interop
             int prefmaxlen,
             out int entriesread,
             out int totalentries,
-            ref int resume_handle);
+            ref IntPtr resume_handle);
 
         [DllImport("netapi32.dll", CharSet = CharSet.Unicode)]
         public static extern int NetUserGetLocalGroups(
